research/requirements.txt
# Anthropic (research_agent.py, research_agent_langgraph.py original)
anthropic>=0.7.0

# OpenAI (research_agent_langgraph.py, research_agent_openai_sdk.py)
openai>=1.0.0
openai-agents>=0.0.1

# Graph-based orchestration (research_agent_langgraph.py)
langgraph>=0.2.0

# Web search
requests>=2.31.0

# Environment / utilities
python-dotenv>=1.0.0
typing_extensions>=4.0.0
